Moved deployment task into its own subtask
