+++ package.json
@@ -24,6 +24,7 @@
     "ethereum-waffle": "^3.4.0",
     "ethers": "^5.4.1",
     "ts-node": "^10.1.0",
+    "tsconfig-paths": "^3.10.1",
     "typescript": "^4.3.5"
   }
 }
